skills/lovecraft/pyproject.toml
[project]
name = "lovecraft-skills"
version = "0.1.0"
description = "CLI helpers for evidence synthesis, date-window clustering, and nested-report outlines (skills toolkit)."
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
lovecraft-tools = "lovecraft_tools.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lovecraft_tools"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
